Strongly type Modal components
