+++ packages/styled-components/src/ComponentModules/Modal/Container.tsx
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+export const Container = styled.div`
+  white-space: normal;
+  cursor: auto;
+  width: 100%;
+  position: relative;
+  display: inline-block;
+  vertical-align: middle;
+  margin: 0 auto;
+  text-align: left;
+`;
