+++ packages/styled-components/src/ComponentModules/Modal/Header.tsx
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const Header = styled.div`
+  display: flex;
+  align-items: flex-start;
+`;
