+++ packages/styled-components/src/ComponentModules/Modal/Modal.tsx
@@ -1,12 +1,14 @@
 import React, { useEffect, useRef } from 'react';
 import { replaceLineBreaks } from '@doctorlink/traversal-core';
 import { defaultModalActions, defaultModalComponents } from './defaults';
+import { ModalModel } from '@doctorlink/traversal-core';
+import { ModalCallbacks } from './ModalCallbacks';
+import { ModalComponents } from './ModalComponents';
 
 export const Modal: React.FC<{
-  modal: any;
-  actions?: any;
-  components?: any;
-  children?: any;
+  modal: ModalModel | null;
+  actions?: ModalCallbacks;
+  components?: ModalComponents;
 }> = ({
   modal,
   children,
@@ -14,20 +16,20 @@ export const Modal: React.FC<{
   components = defaultModalComponents,
 }) => {
   const comps = { ...defaultModalComponents, ...components };
-  const ref = useRef<HTMLElement>();
+  const ref = useRef<HTMLDivElement>(null);
 
   useEffect(() => {
-    const handleClickOutside = (event: any) => {
+    const handleClickOutside = (event: MouseEvent) => {
       if (
         ref &&
         ref.current &&
-        !ref.current.contains(event.target) &&
+        !ref.current.contains(event.target as HTMLElement) &&
         window.innerWidth - event.screenX > 20
       )
         actions.close();
     };
 
-    const handleKeydown = (event: any) => {
+    const handleKeydown = (event: KeyboardEvent) => {
       if (event.keyCode === 27) actions.close();
     };
     document.addEventListener('mousedown', handleClickOutside);
@@ -49,7 +51,7 @@ export const Modal: React.FC<{
         <comps.TransparentCurtain key="curtain">
           <comps.Wrap>
             <comps.Container>
-              <comps.Modal ref={ref}>
+              <comps.ModalWindow ref={ref}>
                 <comps.Header>
                   <comps.Title>{modal.title}</comps.Title>
                   <comps.Close onClick={actions.close} />
@@ -62,7 +64,7 @@ export const Modal: React.FC<{
                   />
                 )}
                 {!modal.content && <comps.Body>{children}</comps.Body>}
-              </comps.Modal>
+              </comps.ModalWindow>
             </comps.Container>
           </comps.Wrap>
         </comps.TransparentCurtain>,
