+++ packages/styled-components/src/ComponentModules/Modal/Title.tsx
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const Title = styled.h3`
+  flex: 1;
+  margin: 0 0 20px;
+`;
