+++ packages/styled-components/src/Components/DelayExit/DelayExit.tsx
@@ -1,22 +1,11 @@
 import React from 'react';
 import { motion } from 'framer-motion';
 
-// export const DelayExit: React.FC<{ children: any, isMounted: boolean, delayTime: number }> = ({ children, isMounted, delayTime = 600 }) => {
-//   const [shouldRender, setShouldRender] = useState<boolean>(isMounted);
-//   const prevIsMounted = usePrevious<boolean>(isMounted);
+export interface DelayExitProps {
+  delay?: number;
+}
 
-//   useEffect(() => {
-//     if (prevIsMounted && !isMounted) {
-//       setTimeout(() => setShouldRender(false), delayTime);
-//     } else if (!prevIsMounted && isMounted) {
-//       setShouldRender(true);
-//     }
-//   }, [prevIsMounted, isMounted, setShouldRender])
-
-//   return shouldRender ? (<>{children}</>) : null;
-// }
-
-const DelayExit: React.FC<{ delay?: number }> = ({ children, delay = 0.3 }) => {
+const DelayExit: React.FC<DelayExitProps> = ({ children, delay = 0.3 }) => {
   return (
     <motion.div
       variants={{
