+++ packages/traversal-core/src/Models/index.ts
@@ -2,6 +2,7 @@ export * from './Traversal';
 export * from './Summary';
 export * from './Conclusion';
 export * from './HealthAssessment';
+export * from './Modal';
 export * from './State';
 export * from './Service';
 export * from './Utility';
